move the util, deprecate the old location
