+++ composer/utils/__init__.py
@@ -20,7 +20,7 @@ from composer.utils.import_helpers import MissingConditionalImportError, import_
 from composer.utils.inference import ExportFormat, Transform, export_for_inference, export_with_logger, quantize_dynamic
 from composer.utils.iter_helpers import IteratorFileStream, ensure_tuple, map_collection
 from composer.utils.misc import (get_free_tcp_port, is_model_deepspeed, is_model_fsdp, is_notebook, model_eval_mode,
-                                 using_torch_2)
+                                 using_torch_2, create_interval_scheduler)
 from composer.utils.object_store import (GCSObjectStore, LibcloudObjectStore, ObjectStore, ObjectStoreTransientError,
                                          OCIObjectStore, S3ObjectStore, SFTPObjectStore, UCObjectStore)
 from composer.utils.retrying import retry
@@ -83,6 +83,7 @@ __all__ = [
     'convert_nested_dict_to_flat_dict',
     'convert_flat_dict_to_nested_dict',
     'using_torch_2',
+    'create_interval_scheduler',
     'EvalClient',
     'LambdaEvalClient',
     'LocalEvalClient',
